登录页使用md5加密传输 在线监控优化

+++ src/main/java/priv/zhou/module/system/monitor/service/impl/SessionServiceImpl.java
@@ -2,7 +2,7 @@ package priv.zhou.module.system.monitor.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.logging.log4j.util.Strings;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.shiro.session.Session;
 import org.apache.shiro.session.UnknownSessionException;
 import org.apache.shiro.session.mgt.DefaultSessionKey;
@@ -10,14 +10,14 @@ import org.apache.shiro.session.mgt.SessionManager;
 import org.apache.shiro.session.mgt.eis.SessionDAO;
 import org.apache.shiro.subject.support.DefaultSubjectContext;
 import org.springframework.stereotype.Service;
-import priv.zhou.common.domain.Result;
 import priv.zhou.common.constant.NULL;
+import priv.zhou.common.domain.Result;
 import priv.zhou.common.enums.ResultEnum;
 import priv.zhou.common.tools.ShiroUtil;
 import priv.zhou.framework.shiro.SyncLoginFilter;
-import priv.zhou.module.system.monitor.domain.dto.OnlineDTO;
-import priv.zhou.module.system.monitor.service.IOnlineService;
-import priv.zhou.module.system.role.domain.dto.RoleDTO;
+import priv.zhou.module.system.monitor.domain.query.SessionQuery;
+import priv.zhou.module.system.monitor.domain.vo.SessionVO;
+import priv.zhou.module.system.monitor.service.ISessionService;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -25,7 +25,7 @@ import java.util.stream.Collectors;
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class OnlineServiceImpl implements IOnlineService {
+public class SessionServiceImpl implements ISessionService {
 
     private final SessionDAO sessionDAO;
 
@@ -35,19 +35,25 @@ public class OnlineServiceImpl implements IOnlineService {
 
 
     @Override
-    public Result<List<OnlineDTO>> list(OnlineDTO onlineDTO) {
+    public Result<List<SessionVO>> list(SessionQuery query) {
         return Result.success(sessionDAO.getActiveSessions().stream()
-                .filter(s -> !(
-                        null == s.getAttribute(DefaultSubjectContext.PRINCIPALS_SESSION_KEY) || // 强制退出，令牌为空
-                                Strings.isNotBlank(onlineDTO.getUsername()) && !((String) s.getAttribute("username")).contains(onlineDTO.getUsername()) ||
-                                null != onlineDTO.getRole() && null != onlineDTO.getRole().getId() && !((RoleDTO) s.getAttribute("role")).getId().equals(onlineDTO.getRole().getId())))
-                .map(s -> new OnlineDTO()
-                        .setHost(s.getHost())
+                .filter(s -> {
+                    if (null == s.getAttribute(DefaultSubjectContext.PRINCIPALS_SESSION_KEY)) {
+                        return false;
+                    } else if (StringUtils.isNotBlank(query.getHost()) && !s.getHost().contains(query.getHost())) {
+                        return false;
+                    } else if (StringUtils.isNotBlank(query.getUsername()) && !((String) s.getAttribute("username")).contains(query.getUsername())) {
+                        return false;
+                    }
+                    return true;
+                })
+                .map(s -> new SessionVO()
                         .setId((String) s.getId())
+                        .setHost(s.getHost())
                         .setLoginTime(s.getStartTimestamp())
                         .setLastAccessTime(s.getLastAccessTime())
                         .setOs((String) s.getAttribute("os"))
-                        .setRole((RoleDTO) s.getAttribute("role"))
+                        .setRoleNames((String) s.getAttribute("roleNames"))
                         .setBrowser((String) s.getAttribute("browser"))
                         .setUsername((String) s.getAttribute("username")))
                 .collect(Collectors.toList())
@@ -61,8 +67,12 @@ public class OnlineServiceImpl implements IOnlineService {
         }
         try {
             Session session = sessionManager.getSession(new DefaultSessionKey(id));
-            syncLoginFilter.remove(session);
+            if (null == session) {
+                return Result.fail(ResultEnum.FAIL_PARAM);
+            }
+            syncLoginFilter.remove((String) session.getAttribute("username"));
             sessionDAO.delete(session);
+            session.stop();
             return Result.success();
         } catch (UnknownSessionException e) {
             return Result.fail(ResultEnum.FAIL_OPERATION, e.getMessage());
